Introducing Timer Functionality
The timer functionality is starting to emerge. The basic functionality being introduced here is a timer for the full time, with periodic events with the amount of time remainig for the 'full time'.
This is bringing a lot of baser classes and supporting roles into existance.

To note here: There are no tests. This is an exploratory phase. The exploration looks to be coming to a soon close. One it closes and the timers feel solid; this will utilize the previously added IMock testing to create tests for everything.
This is a microObject project; this is the cycle of things.

Note: Milliseconds and FromSeconds need to be absorbed into TimeInterval as 'Milliseconds' and be a 'DoubleNumber'. I don't like that name.