RepeatingEventTimer takes TimeInterval
drove creation of a couple classes. A recurring theme at a lot of the
immutability drive for this is around NowAtFirstAccess. This caches on first
access. It existed before; it's just growing to be a collaborator in a larger
number of objects, so it gets called out.

Also added a Seconds and Milliseconds class that gets rid of 'arbitrary' values
for what are Second and Millisecond concepts.