Fixed timer race condition
The TimeLeftTimer (renamed to TimeLeftSingleEventTimer) will sometimes trigger an
update after the actual has fired. This leads to UI glitches as an 'update' fires after the
'done' has displayed. This is bad.
To fix this the RepeatSpecifiedEventTimer (name to be refactored) was created. This
breaks down a duration by a precision and allows a single timer to be both the update and actual. Much cleaner and no race condition.

You should have seen the locking path I was starting to fall down. It was like a nightmare... This requires no locking, and not 3 levels of locking... *shudder*

Tests coming; this was pretty much a bug fix for the UI impacting race condition.