Starting on Desktop App
Starting on the desktop app which is driving the creation of supporting classes. These
classes are things such as number and timeSpan interactions, as well as different
timeinterval creation classes, particularly minutes in this instance.
Some additional clean up around unused timer classes.

A relationship has been identified in the CountDownTimer and this is going to be
extracted.