Commit Point to Show MicroObjects Power
I've made a mistake. I have the decision to be a short break or a long break decided
after the short break timer. This is the entirely wrong place for it. A mistake.
We'll ignore the lack of tests, this was a failing in my thinking about it. Tests MIGHT
have caught it.
What I'm doing with this commit is preserving the state I found the issue in; and the
next commit will show how much, or litte, has to change to correct the flow.