Result of MicroObjects Power
Files didn't migrate well through Visual Studio.
_ShowBreakStart because ShowShortBreakStart
_ShowNextStart got moved from ShortBreak to Session specific location. These files
changed only in location and-or name. Not content. They don't count for the change.

Which leaves Session_TimerFinished and ShortBreak_TimerFinished.
Which each have one line.
I was able to change the logic of showing either the LongBreak or Session into the
correct logic of showing the LongBreak or ShortBreak AS WELL AS move it from the
ShortBreak flow into the Session flow... with... TWO lines changing.

Code Reuse across acrhitectual changes; that's the power. It's powerful.