Pomodoro Timer is functional
While the feature functionality is in place, there's the test pass to do.
I was a bit lazy on testing, granted. The MicroObjects style that I'm VERY familiar with
gives a lot of the advantages of pure TDD, plus some, w/o the actual tests for those
who can see the patterns emerging. It's what I see with long term TDD practitioners
the they can produce high quality code w/o the tests providing the direction.

Time for a test pass as the functionality has stabalized.

The timers aren't appropriate times yet. Looking into how to allow customization to the times.