Adding test project for Desktop App
In the process of adding the test project for the desktop app, I've found I need to
revisit some of the base project types... particularly around the 'not netCore' aspect.
The base projects need to be moved in .Net Standard. I added a second interface
mocks project demonstrating this.

Part of the issue is that I don't have the baser classes in a .net standard project yet.
Once that's updated, it'll resolve this type of issue.